pyproject.toml
[project]
name = "image-to-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "img2pdf>=0.6.1",
    "pillow>=11.3.0",
    "streamlit>=1.46.1",
]
